requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djangorestframework_simplejwt==4.7.2
pytest==6.2.4
